ASC-635 Delete created flavors
This commit adds a teardown step to the pccommon test to delete the
flavors created by the `openstack-ops` deployment. These flavors
conflict with the flavor creation in the `openstack-ansible-ops`
deployment used by other molecule repos.

TODO: It would be preferable to place this code in a pytest
`teardown_module()` method. It is unclear to me how to instanciate a
testinfra host object within such a method to accomplish this.